Separate AFlick pages into subcomponents
Separate AFlick routes into subcomponents:
 - Home
 - Java
 - WebDev
  -DankMemes
 - DistanceLearning
Add NavBar component
