+++ src/src/pages/AFlick/DankMemes.tsx
@@ -1,0 +1,9 @@
+import * as React from "react";
+
+export default function DankMemes(){
+    return (
+        <div className="dank-memes">
+
+        </div>
+    );
+}
