+++ src/src/pages/AFlick/DistanceLearning.tsx
@@ -1,0 +1,9 @@
+import * as React from "react";
+
+export default function DistanceLearning(){
+    return (
+        <div className="distance-learning">
+
+        </div>
+    );
+}
