+++ src/src/pages/AFlick/Java.tsx
@@ -1,0 +1,9 @@
+import * as React from "react";
+
+export default function Java(){
+    return (
+        <div className="java">
+
+        </div>
+    );
+}
