+++ src/src/pages/AFlick/WebDev.tsx
@@ -1,0 +1,9 @@
+import * as React from "react";
+
+export default function WebDev(){
+    return (
+        <div className="web-dev">
+
+        </div>
+    );
+}
